requirements.txt
aiogram==3.2.0
apscheduler==3.10.1
python-dotenv==1.0.0
